Refact: Anomaly Detection - Wrappers, Howtos, RTD #2
