+++ test/howtos/oa/howto_oa_ad_002_oneclasssvm_pa_2d.py
@@ -1,13 +1,14 @@
 ## -------------------------------------------------------------------------------------------------
 ## -- Project : MLPro - The integrative middleware framework for standardized machine learning
 ## -- Package : mlpro_int_scikit_learn
-## -- Module  : howto_oa_ad_005_oneclasssvm_pogo_1d.py
+## -- Module  : howto_oa_ad_002_oneclasssvm_po_2d.py
 ## -------------------------------------------------------------------------------------------------
 ## -- History :
 ## -- yyyy-mm-dd  Ver.      Auth.    Description
 ## -- 2023-04-01  0.0.0     SK       Creation
 ## -- 2023-04-01  1.0.0     SK       First version release
-## -- 2024-05-07  1.0.1     SK       Change in parameter p_outlier_rate
+## -- 2024-04-05  1.0.1     DA       Activated 2D visualization
+## -- 2024-05-07  1.0.2     SK       Change in parameter p_outlier_rate
 ## -------------------------------------------------------------------------------------------------
 
 """
@@ -34,7 +35,7 @@ from mlpro.bf.streams.streams import *
 from mlpro.bf.streams.models import *
 from mlpro.bf.various import Log
 from mlpro.oa.streams import *
-from mlpro_int_sklearn.wrappers.anomalydetectors import WrSklearnOneClassSVM2MLPro
+from mlpro_int_sklearn.wrappers.anomalydetectors.ocsvm import WrSklearnOneClassSVM2MLPro
 
 
 
@@ -49,7 +50,7 @@ class AdScenario4ADsvm (OAScenario):
     def _setup(self, p_mode, p_ada: bool, p_visualize: bool, p_logging):
 
         # 1 Get the native stream from MLPro stream provider
-        mystream = StreamMLProPOutliers( p_functions = ['sin'],
+        mystream = StreamMLProPOutliers( p_functions = ['sin', 'cos'],
                                        p_outlier_rate=0.02,
                                        p_visualize=p_visualize, 
                                        p_logging=p_logging )
@@ -62,7 +63,7 @@ class AdScenario4ADsvm (OAScenario):
                                p_logging=p_logging )
 
         # 3 Initiailise the lof anomaly detctor class
-        anomalydetector = WrSklearnOneClassSVM2MLPro(p_group_anomaly_det=True, p_data_buffer=25, p_delay=3, p_kernel='poly',
+        anomalydetector = WrSklearnOneClassSVM2MLPro(p_group_anomaly_det=False, p_data_buffer=25, p_delay=3, p_kernel='poly',
                                                      p_gamma='scale', p_nu=0.01,p_degree=4,p_coef=0, p_visualize=p_visualize)
 
         # 4 Add anomaly detection task to workflow
@@ -80,7 +81,7 @@ class AdScenario4ADsvm (OAScenario):
 # 1 Preparation of demo/unit test mode
 if __name__ == "__main__":
     # 1.1 Parameters for demo mode
-    cycle_limit = 100
+    cycle_limit = 360
     logging     = Log.C_LOG_ALL
     visualize   = True
     step_rate   = 2
@@ -98,7 +99,7 @@ myscenario = AdScenario4ADsvm( p_mode=Mode.C_MODE_REAL,
                                  p_cycle_limit=cycle_limit,
                                  p_visualize=visualize,
                                  p_logging=logging )
-myscenario.init_plot( p_plot_settings=PlotSettings( p_view = PlotSettings.C_VIEW_ND,
+myscenario.init_plot( p_plot_settings=PlotSettings( p_view = PlotSettings.C_VIEW_2D,
                                                         p_view_autoselect = False,
                                                         p_step_rate = step_rate ) )
