+++ test/howtos/oa/howto_oa_ad_003_oneclasssvm_pa_3d.py
