+++ test/howtos/oa/howto_oa_ad_005_oneclasssvm_paga_1d.py
@@ -1,19 +1,19 @@
 ## -------------------------------------------------------------------------------------------------
 ## -- Project : MLPro - The integrative middleware framework for standardized machine learning
 ## -- Package : mlpro_int_scikit_learn
-## -- Module  : howto_oa_ad_045_if_pogo_1d.py
+## -- Module  : howto_oa_ad_005_oneclasssvm_pogo_1d.py
 ## -------------------------------------------------------------------------------------------------
 ## -- History :
 ## -- yyyy-mm-dd  Ver.      Auth.    Description
-## -- 2024-04-01  0.0.0     SK       Creation
-## -- 2024-04-01  1.0.0     SK       First version release
+## -- 2023-04-01  0.0.0     SK       Creation
+## -- 2023-04-01  1.0.0     SK       First version release
 ## -- 2024-05-07  1.0.1     SK       Change in parameter p_outlier_rate
 ## -------------------------------------------------------------------------------------------------
 
 """
-Ver. 1.0.1 (2024-05-07)
+Ver. 1.0.2 (2024-05-07)
 
-This module demonstrates the use of anomaly detector based on isolation forest algorithm with MLPro.
+This module demonstrates the use of anomaly detector based on one class svm algorithm with MLPro.
 To this regard, a stream of a stream provider is combined with a stream workflow to a stream scenario.
 The workflow consists of a standard task 'Aanomaly Detector'.
 
@@ -26,7 +26,7 @@ You will learn:
 3) How to add a task anomalydetector.
 
 4) How to reuse an anomaly detector algorithm from scikitlearn (https://scikit-learn.org/), specifically
-Isolation Forest
+One Class SVM
 
 """
 
@@ -34,23 +34,23 @@ from mlpro.bf.streams.streams import *
 from mlpro.bf.streams.models import *
 from mlpro.bf.various import Log
 from mlpro.oa.streams import *
-from mlpro_int_sklearn.wrappers.anomalydetectors import WrSklearnIsolationForest2MLPro
+from mlpro_int_sklearn import WrSklearnOneClassSVM2MLPro
 
 
 
 
 ## -------------------------------------------------------------------------------------------------
 ## -------------------------------------------------------------------------------------------------
-class AdScenario4ADif (OAScenario):
+class AdScenario4ADsvm (OAScenario):
 
-    C_NAME = 'AdScenario4ADif'
+    C_NAME = 'AdScenario4ADsvm'
 
 ## -------------------------------------------------------------------------------------------------
     def _setup(self, p_mode, p_ada: bool, p_visualize: bool, p_logging):
 
         # 1 Get the native stream from MLPro stream provider
         mystream = StreamMLProPOutliers( p_functions = ['sin'],
-                                       p_outlier_rate=0.02,
+                                       p_outlier_rate=0.3,
                                        p_visualize=p_visualize, 
                                        p_logging=p_logging )
 
@@ -62,14 +62,15 @@ class AdScenario4ADif (OAScenario):
                                p_logging=p_logging )
 
         # 3 Initiailise the lof anomaly detctor class
-        anomalydetector = WrSklearnIsolationForest2MLPro(p_group_anomaly_det=True, p_estimators=50, p_contamination=0.01, p_data_buffer=25,
-                                                         p_delay=5, p_max_samples=3, p_visualize = p_visualize)
+        anomalydetector = WrSklearnOneClassSVM2MLPro(p_group_anomaly_det=True, p_data_buffer=25, p_delay=3, p_kernel='poly',
+                                                     p_gamma='scale', p_nu=0.01,p_degree=4,p_coef=0, p_visualize=p_visualize)
 
         # 4 Add anomaly detection task to workflow
         workflow.add_task( p_task=anomalydetector )
 
         # 5 Return stream and workflow
         return mystream, workflow
+
 
 
 
@@ -86,18 +87,17 @@ if __name__ == "__main__":
 
 else:
     # 1.2 Parameters for internal unit test
-    cycle_limit = 5
+    cycle_limit = 2
     logging     = Log.C_LOG_NOTHING
     visualize   = False
     step_rate   = 1
 
 
 # 2 Instantiate the stream scenario
-myscenario = AdScenario4ADif( p_mode=Mode.C_MODE_REAL,
+myscenario = AdScenario4ADsvm( p_mode=Mode.C_MODE_REAL,
                                  p_cycle_limit=cycle_limit,
                                  p_visualize=visualize,
                                  p_logging=logging )
-
 myscenario.init_plot( p_plot_settings=PlotSettings( p_view = PlotSettings.C_VIEW_ND,
                                                         p_view_autoselect = False,
                                                         p_step_rate = step_rate ) )
@@ -115,4 +115,3 @@ myscenario.run()
 if __name__ == '__main__':
     input('Press ENTER to exit...')
 
-
