+++ test/howtos/oa/howto_oa_ad_006_oneclasssvm_paga_2d.py
@@ -1,19 +1,20 @@
 ## -------------------------------------------------------------------------------------------------
 ## -- Project : MLPro - The integrative middleware framework for standardized machine learning
 ## -- Package : mlpro_int_scikit_learn
-## -- Module  : howto_oa_ad_022_lof_po_1d.py
+## -- Module  : howto_oa_ad_006_oneclasssvm_po_2d.py
 ## -------------------------------------------------------------------------------------------------
 ## -- History :
 ## -- yyyy-mm-dd  Ver.      Auth.    Description
-## -- 2024-04-01  0.0.0     SK       Creation
-## -- 2024-04-01  1.0.0     SK       First version release
-## -- 2024-05-07  1.0.1     SK       Change in parameter p_outlier_rate
+## -- 2023-04-01  0.0.0     SK       Creation
+## -- 2023-04-01  1.0.0     SK       First version release
+## -- 2024-04-06  1.0.1     DA       Set 2D mode
+## -- 2024-05-07  1.0.2     SK       Change in parameter p_outlier_rate
 ## -------------------------------------------------------------------------------------------------
 
 """
-Ver. 1.0.1 (2024-05-07)
+Ver. 1.0.2 (2024-05-07)
 
-This module demonstrates the use of anomaly detector based on local outlier factor algorithm with MLPro.
+This module demonstrates the use of anomaly detector based on one class svm algorithm with MLPro.
 To this regard, a stream of a stream provider is combined with a stream workflow to a stream scenario.
 The workflow consists of a standard task 'Aanomaly Detector'.
 
@@ -26,7 +27,7 @@ You will learn:
 3) How to add a task anomalydetector.
 
 4) How to reuse an anomaly detector algorithm from scikitlearn (https://scikit-learn.org/), specifically
-Local Outlier Factor
+One Class SVM
 
 """
 
@@ -34,25 +35,25 @@ from mlpro.bf.streams.streams import *
 from mlpro.bf.streams.models import *
 from mlpro.bf.various import Log
 from mlpro.oa.streams import *
-from mlpro_int_sklearn.wrappers.anomalydetectors import WrSklearnLOF2MLPro
+from mlpro_int_sklearn import WrSklearnOneClassSVM2MLPro
 
 
 
 
 ## -------------------------------------------------------------------------------------------------
 ## -------------------------------------------------------------------------------------------------
-class AdScenario4ADlof (OAScenario):
+class AdScenario4ADsvm (OAScenario):
 
-    C_NAME = 'AdScenario4ADlof'
+    C_NAME = 'AdScenario4ADsvm'
 
 ## -------------------------------------------------------------------------------------------------
     def _setup(self, p_mode, p_ada: bool, p_visualize: bool, p_logging):
 
         # 1 Get the native stream from MLPro stream provider
-        mystream = StreamMLProPOutliers( p_functions = ['sin'],
-                                         p_outlier_rate=0.01,
-                                         p_visualize=p_visualize, 
-                                         p_logging=p_logging )
+        mystream = StreamMLProPOutliers( p_functions = ['sin', 'cos'],
+                                       p_outlier_rate = 0.1,
+                                       p_visualize=p_visualize, 
+                                       p_logging=p_logging )
 
         # 2 Creation of a workflow
         workflow = OAWorkflow( p_name='wf1',
@@ -62,13 +63,15 @@ class AdScenario4ADlof (OAScenario):
                                p_logging=p_logging )
 
         # 3 Initiailise the lof anomaly detctor class
-        anomalydetector =WrSklearnLOF2MLPro(p_group_anomaly_det=False, p_neighbours = 3, p_delay=3, p_visualize=p_visualize, p_data_buffer=20)
+        anomalydetector = WrSklearnOneClassSVM2MLPro(p_group_anomaly_det=True, p_data_buffer=25, p_delay=3, p_kernel='poly',
+                                                     p_gamma='scale', p_nu=0.01,p_degree=4,p_coef=0, p_visualize=p_visualize)
 
         # 4 Add anomaly detection task to workflow
         workflow.add_task( p_task=anomalydetector )
 
         # 5 Return stream and workflow
         return mystream, workflow
+
 
 
 
@@ -81,8 +84,8 @@ if __name__ == "__main__":
     cycle_limit = 360
     logging     = Log.C_LOG_ALL
     visualize   = True
-    step_rate   = 1
-  
+    step_rate   = 2
+
 else:
     # 1.2 Parameters for internal unit test
     cycle_limit = 2
@@ -92,12 +95,11 @@ else:
 
 
 # 2 Instantiate the stream scenario
-myscenario = AdScenario4ADlof( p_mode=Mode.C_MODE_REAL,
+myscenario = AdScenario4ADsvm( p_mode=Mode.C_MODE_REAL,
                                  p_cycle_limit=cycle_limit,
                                  p_visualize=visualize,
                                  p_logging=logging )
-
-myscenario.init_plot( p_plot_settings=PlotSettings( p_view = PlotSettings.C_VIEW_ND,
+myscenario.init_plot( p_plot_settings=PlotSettings( p_view = PlotSettings.C_VIEW_2D,
                                                         p_view_autoselect = False,
                                                         p_step_rate = step_rate ) )
 
@@ -114,4 +116,3 @@ myscenario.run()
 if __name__ == '__main__':
     input('Press ENTER to exit...')
 
-
