+++ test/howtos/oa/howto_oa_ad_025_lof_paga_1d.py
@@ -1,18 +1,17 @@
 ## -------------------------------------------------------------------------------------------------
 ## -- Project : MLPro - The integrative middleware framework for standardized machine learning
 ## -- Package : mlpro_int_scikit_learn
-## -- Module  : howto_oa_ad_027_lof_pogo_3d.py
+## -- Module  : howto_oa_ad_025_lof_pogo_1d.py
 ## -------------------------------------------------------------------------------------------------
 ## -- History :
 ## -- yyyy-mm-dd  Ver.      Auth.    Description
 ## -- 2024-04-01  0.0.0     SK       Creation
 ## -- 2024-04-01  1.0.0     SK       First version release
-## -- 2024-04-06  1.0.1     DA       Set 3D mode
-## -- 2024-05-07  1.0.2     SK       Change in parameter p_outlier_rate
+## -- 2024-05-07  1.0.1     SK       Change in parameter p_outlier_rate
 ## -------------------------------------------------------------------------------------------------
 
 """
-Ver. 1.0.2 (2024-05-07)
+Ver. 1.0.1 (2024-05-07)
 
 This module demonstrates the use of anomaly detector based on local outlier factor algorithm with MLPro.
 To this regard, a stream of a stream provider is combined with a stream workflow to a stream scenario.
@@ -35,7 +34,7 @@ from mlpro.bf.streams.streams import *
 from mlpro.bf.streams.models import *
 from mlpro.bf.various import Log
 from mlpro.oa.streams import *
-from mlpro_int_sklearn.wrappers.anomalydetectors import WrSklearnLOF2MLPro
+from mlpro_int_sklearn.wrappers import WrSklearnLOF2MLPro
 
 
 
@@ -50,8 +49,8 @@ class AdScenario4ADlof (OAScenario):
     def _setup(self, p_mode, p_ada: bool, p_visualize: bool, p_logging):
 
         # 1 Get the native stream from MLPro stream provider
-        mystream = StreamMLProPOutliers( p_functions = ['sin', 'cos', 'const'],
-                                       p_outlier_rate=0.02,
+        mystream = StreamMLProPOutliers( p_functions = ['sin'],
+                                       p_outlier_rate=0.3,
                                        p_visualize=p_visualize, 
                                        p_logging=p_logging )
 
@@ -63,7 +62,7 @@ class AdScenario4ADlof (OAScenario):
                                p_logging=p_logging )
 
         # 3 Initiailise the lof anomaly detctor class
-        anomalydetector =WrSklearnLOF2MLPro(p_group_anomaly_det=True, p_neighbours = 3, p_delay=3, p_visualize=p_visualize, p_data_buffer=5)
+        anomalydetector =WrSklearnLOF2MLPro(p_group_anomaly_det=True, p_neighbours = 10, p_delay=5, p_visualize=p_visualize, p_data_buffer=20)
 
         # 4 Add anomaly detection task to workflow
         workflow.add_task( p_task=anomalydetector )
@@ -79,10 +78,10 @@ class AdScenario4ADlof (OAScenario):
 # 1 Preparation of demo/unit test mode
 if __name__ == "__main__":
     # 1.1 Parameters for demo mode
-    cycle_limit = 360
+    cycle_limit = 100
     logging     = Log.C_LOG_ALL
     visualize   = True
-    step_rate   = 2
+    step_rate   = 1
   
 else:
     # 1.2 Parameters for internal unit test
@@ -98,7 +97,7 @@ myscenario = AdScenario4ADlof( p_mode=Mode.C_MODE_REAL,
                                  p_visualize=visualize,
                                  p_logging=logging )
 
-myscenario.init_plot( p_plot_settings=PlotSettings( p_view = PlotSettings.C_VIEW_3D,
+myscenario.init_plot( p_plot_settings=PlotSettings( p_view = PlotSettings.C_VIEW_ND,
                                                         p_view_autoselect = False,
                                                         p_step_rate = step_rate ) )
