+++ test/howtos/oa/howto_oa_ad_028_lof_paga_nd.py
@@ -1,19 +1,19 @@
 ## -------------------------------------------------------------------------------------------------
 ## -- Project : MLPro - The integrative middleware framework for standardized machine learning
 ## -- Package : mlpro_int_scikit_learn
-## -- Module  : howto_oa_ad_001_oneclasssvm_po_1d.py
+## -- Module  : howto_oa_ad_028_lof_pogo_nd.py
 ## -------------------------------------------------------------------------------------------------
 ## -- History :
 ## -- yyyy-mm-dd  Ver.      Auth.    Description
-## -- 2023-04-01  0.0.0     SK       Creation
-## -- 2023-04-01  1.0.0     SK       First version release
+## -- 2024-04-01  0.0.0     SK       Creation
+## -- 2024-04-01  1.0.0     SK       First version release
 ## -- 2024-05-07  1.0.1     SK       Change in parameter p_outlier_rate
 ## -------------------------------------------------------------------------------------------------
 
 """
 Ver. 1.0.1 (2024-05-07)
 
-This module demonstrates the use of anomaly detector based on one class svm algorithm with MLPro.
+This module demonstrates the use of anomaly detector based on local outlier factor algorithm with MLPro.
 To this regard, a stream of a stream provider is combined with a stream workflow to a stream scenario.
 The workflow consists of a standard task 'Aanomaly Detector'.
 
@@ -26,7 +26,7 @@ You will learn:
 3) How to add a task anomalydetector.
 
 4) How to reuse an anomaly detector algorithm from scikitlearn (https://scikit-learn.org/), specifically
-One Class SVM
+Local Outlier Factor
 
 """
 
@@ -34,23 +34,23 @@ from mlpro.bf.streams.streams import *
 from mlpro.bf.streams.models import *
 from mlpro.bf.various import Log
 from mlpro.oa.streams import *
-from mlpro_int_sklearn.wrappers.anomalydetectors import WrSklearnOneClassSVM2MLPro
+from mlpro_int_sklearn.wrappers.anomalydetectors import WrSklearnLOF2MLPro
 
 
 
 
 ## -------------------------------------------------------------------------------------------------
 ## -------------------------------------------------------------------------------------------------
-class AdScenario4ADsvm (OAScenario):
+class AdScenario4ADlof (OAScenario):
 
-    C_NAME = 'AdScenario4ADsvm'
+    C_NAME = 'AdScenario4ADlof'
 
 ## -------------------------------------------------------------------------------------------------
     def _setup(self, p_mode, p_ada: bool, p_visualize: bool, p_logging):
 
         # 1 Get the native stream from MLPro stream provider
-        mystream = StreamMLProPOutliers( p_functions = ['sin'],
-                                       p_outlier_rate=0.02,
+        mystream = StreamMLProPOutliers( p_functions = ['sin', 'cos', 'const'],
+                                       p_outlier_rate=0.1,
                                        p_visualize=p_visualize, 
                                        p_logging=p_logging )
 
@@ -62,15 +62,13 @@ class AdScenario4ADsvm (OAScenario):
                                p_logging=p_logging )
 
         # 3 Initiailise the lof anomaly detctor class
-        anomalydetector = WrSklearnOneClassSVM2MLPro(p_group_anomaly_det=False, p_data_buffer=25, p_delay=3, p_kernel='poly',
-                                                     p_gamma='scale', p_nu=0.01,p_degree=4,p_coef=0, p_visualize=p_visualize)
+        anomalydetector =WrSklearnLOF2MLPro(p_group_anomaly_det=True, p_neighbours = 3, p_delay=3, p_visualize=p_visualize, p_data_buffer=5)
 
         # 4 Add anomaly detection task to workflow
         workflow.add_task( p_task=anomalydetector )
 
         # 5 Return stream and workflow
         return mystream, workflow
-
 
 
 
@@ -83,8 +81,8 @@ if __name__ == "__main__":
     cycle_limit = 360
     logging     = Log.C_LOG_ALL
     visualize   = True
-    step_rate   = 2
-
+    step_rate   = 1
+  
 else:
     # 1.2 Parameters for internal unit test
     cycle_limit = 2
@@ -94,10 +92,11 @@ else:
 
 
 # 2 Instantiate the stream scenario
-myscenario = AdScenario4ADsvm( p_mode=Mode.C_MODE_REAL,
+myscenario = AdScenario4ADlof( p_mode=Mode.C_MODE_REAL,
                                  p_cycle_limit=cycle_limit,
                                  p_visualize=visualize,
                                  p_logging=logging )
+
 myscenario.init_plot( p_plot_settings=PlotSettings( p_view = PlotSettings.C_VIEW_ND,
                                                         p_view_autoselect = False,
                                                         p_step_rate = step_rate ) )
@@ -115,3 +114,4 @@ myscenario.run()
 if __name__ == '__main__':
     input('Press ENTER to exit...')
 
+
