+++ test/howtos/oa/howto_oa_ad_048_if_paga_nd.py
@@ -1,20 +1,19 @@
 ## -------------------------------------------------------------------------------------------------
 ## -- Project : MLPro - The integrative middleware framework for standardized machine learning
 ## -- Package : mlpro_int_scikit_learn
-## -- Module  : howto_oa_ad_028_lof_pogo_nd.py
+## -- Module  : howto_oa_ad_048_if_pogo_nd.py
 ## -------------------------------------------------------------------------------------------------
 ## -- History :
 ## -- yyyy-mm-dd  Ver.      Auth.    Description
-## -- 2024-04-01  0.0.0     SK       Creation
-## -- 2024-04-01  1.0.0     SK       First version release
-## -- 2024-04-06  1.0.1     DA       Set 2D mode
-## -- 2024-05-07  1.0.2     SK       Change in parameter p_outlier_rate
+## -- 2024-04-01  0.0.0     SY       Creation
+## -- 2024-04-01  1.0.0     SY       First version release
+## -- 2024-05-07  1.0.1     SK       Change in parameter p_outlier_rate
 ## -------------------------------------------------------------------------------------------------
 
 """
-Ver. 1.0.2 (2024-05-07)
+Ver. 1.0.1 (2024-05-07)
 
-This module demonstrates the use of anomaly detector based on local outlier factor algorithm with MLPro.
+This module demonstrates the use of anomaly detector based on isolation forest algorithm with MLPro.
 To this regard, a stream of a stream provider is combined with a stream workflow to a stream scenario.
 The workflow consists of a standard task 'Aanomaly Detector'.
 
@@ -27,7 +26,7 @@ You will learn:
 3) How to add a task anomalydetector.
 
 4) How to reuse an anomaly detector algorithm from scikitlearn (https://scikit-learn.org/), specifically
-Local Outlier Factor
+Isolation Forest
 
 """
 
@@ -35,23 +34,23 @@ from mlpro.bf.streams.streams import *
 from mlpro.bf.streams.models import *
 from mlpro.bf.various import Log
 from mlpro.oa.streams import *
-from mlpro_int_sklearn.wrappers.anomalydetectors import WrSklearnLOF2MLPro
+from mlpro_int_sklearn.wrappers.anomalydetectors import WrSklearnIsolationForest2MLPro
 
 
 
 
 ## -------------------------------------------------------------------------------------------------
 ## -------------------------------------------------------------------------------------------------
-class AdScenario4ADlof (OAScenario):
+class AdScenario4ADif (OAScenario):
 
-    C_NAME = 'AdScenario4ADlof'
+    C_NAME = 'AdScenario4ADif'
 
 ## -------------------------------------------------------------------------------------------------
     def _setup(self, p_mode, p_ada: bool, p_visualize: bool, p_logging):
 
         # 1 Get the native stream from MLPro stream provider
-        mystream = StreamMLProPOutliers( p_functions = ['sin', 'cos'],
-                                       p_outlier_rate= 0.02,
+        mystream = StreamMLProPOutliers( p_functions = ['sin', 'cos', 'const'],
+                                       p_outlier_rate=0.1,
                                        p_visualize=p_visualize, 
                                        p_logging=p_logging )
 
@@ -62,27 +61,14 @@ class AdScenario4ADlof (OAScenario):
                                p_visualize=p_visualize, 
                                p_logging=p_logging )
 
-        # 3 Initialize a Boundarydetector task
-        task_bd = BoundaryDetector( p_name='Demo Boundary Detector', 
-                                                 p_ada=p_ada, 
-                                                 p_visualize=p_visualize,
-                                                 p_logging=p_logging )
-        
-        # 4 Initialize a Normalizer task
-        task_norm = NormalizerMinMax( p_name='Demo MinMax Normalizer', 
-                                                 p_ada=p_ada, 
-                                                 p_visualize=p_visualize,
-                                                 p_logging=p_logging)
+        # 3 Initiailise the lof anomaly detctor class
+        anomalydetector = WrSklearnIsolationForest2MLPro(p_group_anomaly_det=True, p_estimators=50, p_contamination=0.01, p_data_buffer=25,
+                                                         p_delay=5, p_max_samples=3, p_visualize = p_visualize)
 
-        # 5 Initiailise the lof anomaly detctor class
-        anomalydetector =WrSklearnLOF2MLPro(p_group_anomaly_det=True, p_neighbours = 3, p_delay=3, p_visualize=p_visualize, p_data_buffer=5)
+        # 4 Add anomaly detection task to workflow
+        workflow.add_task( p_task=anomalydetector )
 
-        # 6 Addition of the task to the workflow
-        workflow.add_task(p_task = task_bd)
-        workflow.add_task(p_task = task_norm, p_pred_tasks=[task_bd])
-        workflow.add_task(p_task=anomalydetector, p_pred_tasks=[task_norm] )
-
-        # 7 Return stream and workflow
+        # 5 Return stream and workflow
         return mystream, workflow
 
 
@@ -96,18 +82,18 @@ if __name__ == "__main__":
     cycle_limit = 100
     logging     = Log.C_LOG_ALL
     visualize   = True
-    step_rate   = 1
-  
+    step_rate   = 2
+
 else:
     # 1.2 Parameters for internal unit test
-    cycle_limit = 2
+    cycle_limit = 5
     logging     = Log.C_LOG_NOTHING
     visualize   = False
     step_rate   = 1
 
 
 # 2 Instantiate the stream scenario
-myscenario = AdScenario4ADlof( p_mode=Mode.C_MODE_REAL,
+myscenario = AdScenario4ADif( p_mode=Mode.C_MODE_REAL,
                                  p_cycle_limit=cycle_limit,
                                  p_visualize=visualize,
                                  p_logging=logging )
